Renaming it back to cluster_metadata_provider.go
